backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
psycopg[binary]==3.3.2
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.12
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
